New prototype version
Expanding test coverage. Bug fixes (climates now getting allocated to water year instead of calendar year, fixed errors in calculation of the complex EWRs in the Murray that have two part requirements [409025 OB2 and OB3], fixed errors in the calculation of the Macquarie EWRs which have the multi flow and simultaneous flow requirements [EWRs at the 421090 gauge]. Logic fixes to the summary statistics (statistics requiring averaging in the summary table are now getting calculated based on all events, not the aggregated yearly statistics).